requirements.txt
Pillow-SIMD==6.0.0.post0
argparse==1.1
